fix: use raw img in iso card details
The NuxtImg module does not create the requested sizes on generate when a v-if
conditional is used. [1]

With v-show it works, but then the 'Back' button of card is not properly
layouted. For now instead just use a raw img tag for the label. It has a small
size anyway.

[1] https://github.com/nuxt/image/issues/215#issuecomment-1804847465