feat: download full ISO images by default
The minimal image is leaner and user can in theory install missing software
later on via Pamac but it might still miss some crucial system components
making it more difficult for novice users.

Because of that for now we switch to the full image as standard download and
only provide the minimal image link in the details view.