fix: force creation of NuxtImg files
The images for the phones were not generated because they are only reached by
a device selection at runtime.

This is a similar problem as with 4b71d9ebfc692 and has been noted in an issue
ticket upstream at [1].

To still create the files wheng generating the static page, we simply add the
missing ones to the ARM download page in a hidden state.

[1] https://github.com/nuxt/image/issues/215#issuecomment-1804847465