feat: add device query to ARM downloads
This allows to link directly to the ISO of a specific ARM device.

Note that we need to force to render the cards now client-side because
otherwise the route is not always correctly updated. Since for showing the
cards in a correct way required JS to begin with though because of the
dropdown menu, we don't lose functionality.